backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "belvo-python>=0.8.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.35.0",
    "requests>=2.32.0",
    "pluggy-sdk>=1.0.0.post53",
]
